requirements.txt
# =============================================================================
# REBATE MANAGEMENT BACKEND - DEPENDENCIAS IMPLEMENTADAS
# =============================================================================

# =============================================================================
# FRAMEWORK WEB Y SERVIDOR
# =============================================================================
fastapi==0.116.1
uvicorn[standard]==0.35.0
starlette==0.47.2

# =============================================================================
# VALIDACIÓN Y SERIALIZACIÓN
# =============================================================================
pydantic==2.9.2
pydantic-settings==2.10.1
pydantic_core==2.23.4
python-multipart==0.0.20
email-validator==2.2.0

# =============================================================================
# BASE DE DATOS Y ORM
# =============================================================================
SQLAlchemy==2.0.21
psycopg2-binary==2.9.10
asyncpg==0.29.0
alembic==1.16.4
greenlet==3.2.4

# =============================================================================
# SEGURIDAD Y AUTENTICACIÓN
# =============================================================================
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
python-jose[cryptography]==3.5.0
cryptography==45.0.6
Authlib==1.6.1
PyJWT>=2.8.0

# =============================================================================
# RATE LIMITING Y CACHING
# =============================================================================
slowapi==0.1.9
fastapi-limiter==0.1.6
redis==6.4.0

# =============================================================================
# UTILIDADES Y CONFIGURACIÓN
# =============================================================================
python-dotenv==1.1.1
bleach==6.2.0
tenacity==9.1.2
watchfiles==1.1.0
pandas==2.3.0
python-dateutil==2.8.2
XlsxWriter==3.1.9
openpyxl==3.1.2

# =============================================================================
# GOOGLE CLOUD BIGQUERY
# =============================================================================
google-cloud-bigquery==3.17.2
google-auth==2.28.1
google-cloud-core==2.4.1
google-api-core==2.19.0
google-cloud-storage==2.16.0

# =============================================================================
# AZURE CLOUD
# =============================================================================
azure-storage-blob==12.19.0

# =============================================================================
# DEPENDENCIAS DE DESARROLLO Y TESTING
# =============================================================================
pytest==7.4.3
pytest-cov==6.2.1
pytest-bdd==7.0.1
pytest-mock==3.14.1
pytest-xdist==3.8.0
coverage==7.10.4
httpx==0.28.1
ruff==0.12.8
pytest-asyncio==0.23.6

# =============================================================================
# DEPENDENCIAS ADICIONALES IDENTIFICADAS
# =============================================================================
# Para compatibilidad con Python < 3.9
typing-extensions==4.12.2
# Para manejo de fechas y tiempo
arrow==1.3.0
# Para validación de datos
marshmallow==3.23.1
# Para manejo de archivos y rutas
pathlib2==2.3.7
